Backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-multipart
python-dotenv
pydantic
typing-extensions